LidarLite: Code cleanup, prepare for PWM version of the driver
